+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @unifil/react-native
 
+## 1.0.6
+
+### Patch Changes
+
+- adjustment customTab
+
 ## 1.0.5
 
 ### Patch Changes